add: PageHero to Approach page and remove from temp location for rendering
