+++ src/pages/approach/index.tsx
@@ -1,11 +1,23 @@
 import { NextPage } from 'next';
-
 import { Layout } from '~/components';
+
+import { PageHero } from '~/components/shared';
+
+const imgSource = {
+  desktop: '/shared/approach-hero-xlarge.jpg',
+  tablet: '/shared/approach-hero-medium.jpg',
+  xl: '/shared/approach-hero-large.jpg',
+  xs: '/shared/approach-hero-small.jpg',
+};
 
 const ApproachIndexPage: NextPage = () => {
   return (
     <Layout>
-      <h1>Approach challenges require unique solutions.</h1>
+      <PageHero
+        alt="person staring at computer"
+        imgSource={imgSource}
+        text="People motivate what we do and how we do it"
+      />
     </Layout>
   );
 };
